rendering/requirements.txt
Django==3.1.4
django-concurrency
django-cors-headers
django-admin-list-filter-dropdown==1.0.2
django-environ
social-auth-app-django
wcmatch
Pillow
gunicorn
pika
